method for getting requestStatusId when request creating

+++ BPMFlow.Application/Services/RequestStatusTransitionService.cs
@@ -39,7 +39,7 @@ public class RequestStatusTransitionService : IRequestStatusTransitionService
             var currentEmployeeRoleInRequest = await _employeeRoleService.GetRoleInRequest(employee.Id, objectRequest.ObjectId, objectRequest.Code);
             var currentEmployeeRoleInOrgStructure = await _employeeRoleService.GetRoleInOrgStructure(employee.Id, objectRequest.ObjectId);
 
-            var requestStatus = await _unitOfWork.RequestStatusRepository.GetById(objectRequest.RequestStatusId);
+            var requestStatus = await _unitOfWork.RequestStatusRepository.GetById((int)objectRequest.RequestStatusId);
 
             transitions = await _unitOfWork.RequestStatusTransitionRepository.GetAvailableTransition
                                             (x => x.RequestId == requestStatus.RequestId
